+++ rana_qgis_plugin/loader.py
@@ -51,6 +51,7 @@ from rana_qgis_plugin.utils_api import (
     create_folder,
     delete_tenant_project_directory,
     delete_tenant_project_file,
+    get_process_id_for_tag,
     get_tenant_details,
     get_tenant_file_descriptor,
     get_tenant_file_descriptor_view,
@@ -570,12 +571,6 @@ class Loader(QObject):
             if simulation_wizard.exec() == QDialog.Rejected:
                 self.simulation_cancelled.emit()
 
-    def get_process_id_for_tag(self, tag: str) -> Optional[str]:
-        processes = get_tenant_processes(self.communication)
-        for process in processes:
-            if tag in process["tags"]:
-                return process["id"]
-
     def start_model_tracker_process(
         self,
         project,
@@ -583,7 +578,7 @@ class Loader(QObject):
         revision_id: int,
         inherit_from_previous_revision: bool = True,
     ):
-        track_process = self.get_process_id_for_tag("model_tracker")
+        track_process = get_process_id_for_tag(self.communication, "model_tracker")
         if track_process is None:
             self.communication.log_err("No model tracker available")
             return
@@ -602,7 +597,7 @@ class Loader(QObject):
 
     def start_simulation_tracker_process(self, project, file, simulations):
         # Find the simulation tracker processes
-        track_process = self.get_process_id_for_tag("simulation_tracker")
+        track_process = get_process_id_for_tag(self.communication, "simulation_tracker")
 
         if track_process is None:
             self.communication.log_err("No simulation tracker available")
@@ -1308,3 +1303,20 @@ class Loader(QObject):
         self.project_job_monitor.job_updated.connect(self.project_job_updated)
         self.project_job_monitor.failed.connect(self.communication.show_warn)
         self.project_job_monitor.start()
+
+    @pyqtSlot(int)
+    def cancel_simulation(self, simulation_pk):
+        confirm_cancel = QMessageBox.warning(
+            None,
+            "Cancel Simulation",
+            "Are you sure you want to cancel the simulation?",
+            QMessageBox.StandardButton.Yes,
+            QMessageBox.StandardButton.No,
+        )
+        if confirm_cancel == QMessageBox.StandardButton.Yes:
+            tc = ThreediCalls(get_threedi_api())
+            tc.fetch_simulation_status(simulation_pk)
+            try:
+                tc.create_simulation_action(simulation_pk, name="shutdown")
+            except ApiException as e:
+                self.communication.show_error(f"Could not cancel simulation")